+++ src/org/jitsi/impl/neomedia/jmfext/media/protocol/directshow/DataSource.java
@@ -35,8 +35,8 @@ public class DataSource
      * The map of DirectShow pixel formats to FFmpeg pixel formats which allows
      * converting between the two.
      */
-    private static final long[] DS_TO_FFMPEG_PIX_FMTS
-        = new long[]
+    private static final int[] DS_TO_FFMPEG_PIX_FMTS
+        = new int[]
                 {
                     DSFormat.RGB24,
                     FFmpeg.PIX_FMT_RGB24,
@@ -71,7 +71,7 @@ public class DataSource
      * @param ffmpegPixFmt FFmpeg format
      * @return the DirectShow pixel format matching the specified FFmpeg format
      */
-    public static long getDSPixFmt(int ffmpegPixFmt)
+    public static int getDSPixFmt(int ffmpegPixFmt)
     {
         for (int i = 0; i < DS_TO_FFMPEG_PIX_FMTS.length; i += 2)
             if (DS_TO_FFMPEG_PIX_FMTS[i + 1] == ffmpegPixFmt)
@@ -87,7 +87,7 @@ public class DataSource
      * FFmpeg pixel format of
      * @return the FFmpeg pixel format matching the specified DirectShow pixel
      */
-    public static long getFFmpegPixFmt(long dsPixFmt)
+    public static int getFFmpegPixFmt(int dsPixFmt)
     {
         for (int i = 0; i < DS_TO_FFMPEG_PIX_FMTS.length; i += 2)
             if (DS_TO_FFMPEG_PIX_FMTS[i] == dsPixFmt)
@@ -381,14 +381,14 @@ public class DataSource
         {
             Dimension size
                 = new Dimension(deviceFmt.getWidth(), deviceFmt.getHeight());
-            long devicePixFmt = deviceFmt.getPixelFormat();
-            int pixFmt = (int) getFFmpegPixFmt(devicePixFmt);
+            int devicePixFmt = deviceFmt.getPixelFormat();
+            int pixFmt = getFFmpegPixFmt(devicePixFmt);
 
             fmts.add(
                     new AVFrameFormat(
                             size,
                             Format.NOT_SPECIFIED,
-                            pixFmt, (int) devicePixFmt));
+                            pixFmt, devicePixFmt));
         }
         return fmts.toArray(new Format[fmts.size()]);
     }
